+++ netpyne/analysis/__init__.py
@@ -36,7 +36,7 @@ from .network import plotConn, _plotConnCalculateFromSim, _plotConnCalculateFrom
 # -------------------------------------------------------------------------------------------------------------------
 # Import spike-related functions
 # -------------------------------------------------------------------------------------------------------------------
-from .spikes import calculateRate, plotRates, plotSyncs, plotRaster, plotSpikeHist, plotSpikeStats, plotRatePSD, popAvgRates
+from .spikes import calculateRate, plotRates, plotSyncs, plotRaster, plotSpikeHist, plotSpikeStats, plotRatePSD, plotRateSpectrogram, popAvgRates
 
 
 # -------------------------------------------------------------------------------------------------------------------